Drill Timer countdown
Improved to auto drill Timmer, to count down, and display how many seconds are left until the next move is generated. This also inherently improved the pause functionality, as when you resume, it does not start the timer over again but instead resumes with the amount of time that was left when you paused.