Music Controls and BPM Filtering
Starting adding song start time trims , as well as calculating the BPMs for some of the songs in the json song URL dataset. Added play controls to allow pausing music, skipping to the next song, and setting the desired min and max BPM we want for the songs that get selected. This feature only works for songs that have a BPM added in the dataset for that song. Any song will still be picked if no songs are found with the desired BPMs