Dance Move Data Object
Upgraded the Dance move string array, into a array of dance move objects, so more information can be held about a dance move such as is rhythm. Added a function that can count how many beets a dance move has using its Rhythm. Also added a function that converts the index to the same 1A. 1B. format in the syllabus