Count Speach attempt
Tried adding feature you count using the SPeach API. It sortof works, but the playbeac speed is not reliable enough. takes too much time to lead the speach. over all it talks way to slow and is not on time as much as i want. Back logs of speach also Que up making it say random things when moving on two new moves