Audio Clip Based Rhythm Count
The speech count was not working, so i recoded all the counting elements my self, and now have an audio cliv version that does a much better job a being timing precise.  All audio clip are preloaded so there is no buffer time. Can adjust the BPM that the counts are narated out