Counting Wait For first one with music
First implementation for calcualting and waining for the first one before counting when music is playing. Currently just assumes the first one is when the music Start time is. This was originally just a time code that trimmed the empty space before the music started, so im may not be exactly one... but it could be